Rename `model_fixtures` to `semantic_manifest_fixtures`
