+++ tests/conftest.py
@@ -1,2 +1,2 @@
 # These imports are required to properly set up pytest fixtures.
-from tests.fixtures.model_fixtures import *  # noqa: F401, F403
+from tests.fixtures.semantic_manifest_fixtures import *  # noqa: F401, F403
