+++ tests/fixtures/semantic_manifest_fixtures.py
